+++ CHANGELOG.md
@@ -2,13 +2,16 @@
 
 ## 4.3.0 - Unreleased
 ### Added
-- Added a new `ssiEnabled` config setting that enables Blitz to include templates using Server Side Includes (SSI), which must be enabled on the server.
+- Added the `ssiEnabled` config setting that enables Blitz to include templates using Server Side Includes (SSI), which must be enabled on the server.
+- Added the `esiEnabled` config setting that enables Blitz to include templates using Edge Side Includes (ESI), which must be enabled on the server or CDN.
 - Added the `craft.blitz.staticInclude()` template variable, that includes a template using SSI if enabled, otherwise via an AJAX request. The `staticInclude()` method returns a cached result (if one exists).
 - Added the `craft.blitz.dynamicInclude()` template variable, that includes a template via an AJAX request. The `dynamicInclude()` method always returns a freshly rendered template.
+- Added the `craft.blitz.fetchUri()` template variable, that fetches a URI via an AJAX request. Whether the URI response is cached or not is determined by the URI patterns in the plugin settings.
 - Added the `blitz/templates/static-include` and `blitz/templates/dynamic-include` controller actions.
 
 ### Deprecated
 - Deprecated the `craft.blitz.getTemplate()` template variable. Use `craft.blitz.staticInclude()` or `craft.blitz.dynamicInclude()` instead.
+- Deprecated the `craft.blitz.getUri()` template variable. Use `craft.blitz.fetchUri()` instead.
 - Deprecated the `blitz/templates/get` controller action. Use `blitz/templates/static-include` or `blitz/templates/dynamic-include` instead.
 
 ## 4.2.3 - 2022-10-19